refactor: move rpc into a crate
