+++ crates/rpc/src/pathfinder.rs
