+++ crates/rpc/src/v02/method/add_declare_transaction.rs
