+++ crates/rpc/src/v02/method/get_block_transaction_count.rs
