+++ crates/rpc/src/v02/method/get_class.rs
