+++ crates/rpc/src/v02/method/get_class_hash_at.rs
