+++ crates/rpc/src/v02/method/get_transaction_by_block_id_and_index.rs
