+++ crates/rpc/src/v02/method/get_transaction_by_hash.rs
