+++ crates/rpc/src/v02/types/class.rs
